requirements.txt
django~=4.1.7
djangorestframework~=3.14.0
pillow
psycopg2-binary
drf-yasg
django-filter~=22.1
dj-rest-auth
python-decouple
requests~=2.28.2